app/requirements.txt
decorator==4.0.9
Django==1.9.2
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
ipython==4.1.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.1.0
